Move verifyEmail to pages/ using getServerSideProps
